+++ open-bike-game/src/app/app.module.ts
@@ -5,9 +5,9 @@ import { FormsModule } from '@angular/forms';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { EditRaceComponent } from './edit-race/edit-race.component';
-import { MainMenuComponent } from './main-menu/main-menu.component';
-import { RaceViewComponent } from './race-view/race-view.component';
-import { RaceComponent } from './race/race.component';
+import { MainMenuComponent } from './components/main-menu/main-menu.component';
+import { RaceViewComponent } from './components/race-view/race-view.component';
+import { RaceComponent } from './components/race/race.component';
 
 @NgModule({
   declarations: [
