+++ open-bike-game/src/app/components/main-menu/main-menu.component.spec.ts
