+++ open-bike-game/src/app/components/race-view/race-view.component.scss
