--- open-bike-game/src/app/race/race.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-race',
-  templateUrl: './race.component.html',
-  styleUrls: ['./race.component.scss']
-})
-export class RaceComponent {
-
-}
